Fix issues related to template collections and add notes about dev server mode

+++ packages/@tinacms/app/src/index.ts
@@ -109,8 +109,24 @@ vite.svg`
     },
     logLevel: 'silent',
   }
-  await build(config)
-  await fs.rmSync(out)
+  if (true) {
+    await build(config)
+    await fs.rmSync(out)
+  } else {
+    /**
+     * Uncomment to run the dev server
+     * Note that this assumes the outputFile is 'admin'
+     * And will run into port issues when the build server
+     * restart itself
+     */
+    // const indexDev = await fs
+    //   .readFileSync(path.join(root, 'index.dev.html'))
+    //   .toString()
+    // await fs.writeFileSync(path.join(outDir, 'index.html'), indexDev)
+    // const server = await createServer(config)
+    // await server.listen()
+    // await server.printUrls()
+  }
 }
 
 const loaders: { [ext: string]: Loader } = {
